Creates auth server with password grant (#3)
A username and password can be exchanged for a jwt token.
The token contains the user role. The token can then be included
as a header in the format
"Authorization": "Bearer (token)"
The api endpoints can then read the role from the token to
verify the user is allowed to access the resource requested.